+++ src/vault-api/Controller.ts
@@ -4,12 +4,15 @@
 import Client from './client';
 
 import {
-  validateConnectionConfig, validateInsertRecords, validateDetokenizeInput, validateGetByIdInput, validateInitConfig, validateUpsertOptions, validateUpdateInput, validateGetInput,
+  validateConnectionConfig, validateInsertRecords, validateDetokenizeInput, validateGetByIdInput, validateInitConfig, validateUpsertOptions, validateUpdateInput, validateGetInput, validateDeleteInputAndOptions
 } from './utils/validators';
 
 import {
   ContentType,
+  IDeleteInput,
+  IDeleteOptions,
   IGetInput,
+  IGetOptions,
   IInsertOptions,
   IUpdateInput,
   IUpdateOptions,
@@ -40,6 +43,7 @@ import {
 import { isTokenValid } from './utils/jwt-utils';
 import SKYFLOW_ERROR_CODE from './utils/constants';
 import SkyflowError from './libs/SkyflowError';
+import { deleteRecordsBySkyflowID } from './core/Delete';
 
 class Controller {
   #client: Client;
@@ -148,12 +152,12 @@ class Controller {
     });
   }
 
-  get(getInput: IGetInput) {
+  get(getInput: IGetInput,options:IGetOptions={}) {
     return new Promise((resolve, reject) => {
       try {
         validateInitConfig(this.#client.config)
         printLog(logs.infoLogs.VALIDATE_GET_INPUT, MessageType.LOG);
-        validateGetInput(getInput);
+        validateGetInput(getInput,options);
         printLog(parameterizedString(logs.infoLogs.EMIT_REQUEST,
           TYPES.GET),
           MessageType.LOG);
@@ -161,7 +165,8 @@ class Controller {
           fetchRecordsBySkyflowID(
             getInput.records,
             this.#client,
-            res
+            res,
+            options
           ).then(
             (resolvedResult) => {
               printLog(logs.infoLogs.GET_BY_SKYFLOWID_RESOLVED, MessageType.LOG);
@@ -241,6 +246,32 @@ class Controller {
         reject(error);
       }
     });
+  }
+
+  delete(deleteInput: IDeleteInput, options?: IDeleteOptions) {
+    return new Promise((resolve, reject) => {
+      try {
+        validateInitConfig(this.#client.config);
+        validateDeleteInputAndOptions(deleteInput, options);
+        this.getToken().then((authToken) => {
+          deleteRecordsBySkyflowID(deleteInput.records, options = {}, this.#client, authToken)
+            .then((response) => {
+              printLog(logs.infoLogs.DELETE_REQUEST_RESOLVED, MessageType.LOG);
+              resolve(response);
+            }).catch((error) => {
+              printLog(logs.errorLogs.DELETE_REQUEST_REJECTED, MessageType.ERROR);
+              reject(error);
+            });
+        }).catch((err) => {
+          reject(err);
+        })
+
+      } catch (error) {
+        if (error instanceof Error)
+          printLog(error.message, MessageType.ERROR);
+        reject(error);
+      }
+    })
   }
 
   insertData(records, options) {