+++ src/vault-api/utils/common/index.ts
@@ -103,7 +103,7 @@ export interface IDetokenizeInput {
  */
 export interface ISkyflowIdRecord {
   ids?: string[];
-  redaction: RedactionType;
+  redaction?: RedactionType;
   table: string;
   columnName?: string;
   columnValues?: string[];
@@ -224,6 +224,23 @@ export interface IUpdateInput{
  */
 export interface IUpdateOptions{
   tokens: boolean
+}
+
+export interface IGetOptions{
+  tokens?: boolean
+  encodeURI?: boolean
+}
+export interface IDeleteRecord {
+  id: string;
+  table: string;
+}
+
+export interface IDeleteInput {
+  records: IDeleteRecord[];
+}
+
+export interface IDeleteOptions {
+
 }
 
 export const SDK_METRICS_HEADER_KEY = "sky-metadata";